Avoid functions due to global install conflicts
