+++ src/functions.php
@@ -1,13 +1,6 @@
 <?php
 namespace GuzzleHttp;
 
-use GuzzleHttp\Exception\InvalidArgumentException;
-use GuzzleHttp\Handler\CurlHandler;
-use GuzzleHttp\Handler\CurlMultiHandler;
-use GuzzleHttp\Handler\Proxy;
-use GuzzleHttp\Handler\StreamHandler;
-use Psr\Http\Message\UriInterface;
-
 /**
  * Debug function used to describe the provided value type and class.
  *
@@ -16,17 +9,7 @@ use Psr\Http\Message\UriInterface;
  */
 function describe_type($input): string
 {
-    switch (\gettype($input)) {
-        case 'object':
-            return 'object(' . \get_class($input) . ')';
-        case 'array':
-            return 'array(' . \count($input) . ')';
-        default:
-            \ob_start();
-            \var_dump($input);
-            // normalize float vs double
-            return \str_replace('double(', 'float(', \rtrim(\ob_get_clean()));
-    }
+    return Utils::describeType($input);
 }
 
 /**
@@ -37,16 +20,7 @@ function describe_type($input): string
  */
 function headers_from_lines($lines): array
 {
-    $headers = [];
-
-    foreach ($lines as $line) {
-        $parts = \explode(':', $line, 2);
-        $headers[\trim($parts[0])][] = isset($parts[1])
-            ? \trim($parts[1])
-            : null;
-    }
-
-    return $headers;
+    return Utils::headersFromLines($lines);
 }
 
 /**
@@ -58,13 +32,7 @@ function headers_from_lines($lines): array
  */
 function debug_resource($value = null)
 {
-    if (\is_resource($value)) {
-        return $value;
-    } elseif (\defined('STDOUT')) {
-        return STDOUT;
-    }
-
-    return \fopen('php://output', 'w');
+    return Utils::debugResource($value);
 }
 
 /**
@@ -78,43 +46,15 @@ function debug_resource($value = null)
  */
 function choose_handler(): callable
 {
-    $handler = null;
-    if (\function_exists('curl_multi_exec') && \function_exists('curl_exec')) {
-        $handler = Proxy::wrapSync(new CurlMultiHandler(), new CurlHandler());
-    } elseif (\function_exists('curl_exec')) {
-        $handler = new CurlHandler();
-    } elseif (\function_exists('curl_multi_exec')) {
-        $handler = new CurlMultiHandler();
-    }
-
-    if (\ini_get('allow_url_fopen')) {
-        $handler = $handler
-            ? Proxy::wrapStreaming($handler, new StreamHandler())
-            : new StreamHandler();
-    } elseif (!$handler) {
-        throw new \RuntimeException('GuzzleHttp requires cURL, the '
-            . 'allow_url_fopen ini setting, or a custom HTTP handler.');
-    }
-
-    return $handler;
+    return Utils::chooseHandler();
 }
 
 /**
- * Get the default User-Agent string to use with Guzzle
+ * Get the default User-Agent string to use with Guzzle.
  */
 function default_user_agent(): string
 {
-    static $defaultAgent = '';
-
-    if (!$defaultAgent) {
-        $defaultAgent = 'GuzzleHttp/Guzzle';
-        if (\extension_loaded('curl') && \function_exists('curl_version')) {
-            $defaultAgent .= ' curl/' . \curl_version()['version'];
-        }
-        $defaultAgent .= ' PHP/' . PHP_VERSION;
-    }
-
-    return $defaultAgent;
+    return Utils::defaultUserAgent();
 }
 
 /**
@@ -132,59 +72,7 @@ function default_user_agent(): string
  */
 function default_ca_bundle(): string
 {
-    static $cached = null;
-    static $cafiles = [
-        // Red Hat, CentOS, Fedora (provided by the ca-certificates package)
-        '/etc/pki/tls/certs/ca-bundle.crt',
-        // Ubuntu, Debian (provided by the ca-certificates package)
-        '/etc/ssl/certs/ca-certificates.crt',
-        // FreeBSD (provided by the ca_root_nss package)
-        '/usr/local/share/certs/ca-root-nss.crt',
-        // SLES 12 (provided by the ca-certificates package)
-        '/var/lib/ca-certificates/ca-bundle.pem',
-        // OS X provided by homebrew (using the default path)
-        '/usr/local/etc/openssl/cert.pem',
-        // Google app engine
-        '/etc/ca-certificates.crt',
-        // Windows?
-        'C:\\windows\\system32\\curl-ca-bundle.crt',
-        'C:\\windows\\curl-ca-bundle.crt',
-    ];
-
-    if ($cached) {
-        return $cached;
-    }
-
-    if ($ca = \ini_get('openssl.cafile')) {
-        return $cached = $ca;
-    }
-
-    if ($ca = \ini_get('curl.cainfo')) {
-        return $cached = $ca;
-    }
-
-    foreach ($cafiles as $filename) {
-        if (\file_exists($filename)) {
-            return $cached = $filename;
-        }
-    }
-
-    throw new \RuntimeException(
-        <<< EOT
-No system CA bundle could be found in any of the the common system locations.
-PHP versions earlier than 5.6 are not properly configured to use the system's
-CA bundle by default. In order to verify peer certificates, you will need to
-supply the path on disk to a certificate bundle to the 'verify' request
-option: http://docs.guzzlephp.org/en/latest/clients.html#verify. If you do not
-need a specific certificate bundle, then Mozilla provides a commonly used CA
-bundle which can be downloaded here (provided by the maintainer of cURL):
-https://raw.githubusercontent.com/bagder/ca-bundle/master/ca-bundle.crt. Once
-you have a CA bundle available on disk, you can set the 'openssl.cafile' PHP
-ini setting to point to the path to the file, allowing you to omit the 'verify'
-request option. See http://curl.haxx.se/docs/sslcerts.html for more
-information.
-EOT
-    );
+    return Utils::defaultCaBundle();
 }
 
 /**
@@ -193,12 +81,7 @@ EOT
  */
 function normalize_header_keys(array $headers): array
 {
-    $result = [];
-    foreach (\array_keys($headers) as $key) {
-        $result[\strtolower($key)] = $key;
-    }
-
-    return $result;
+    return Utils::normalizeHeaderKeys($headers);
 }
 
 /**
@@ -217,39 +100,12 @@ function normalize_header_keys(array $headers): array
  *
  * @param string   $host         Host to check against the patterns.
  * @param string[] $noProxyArray An array of host patterns.
+ *
+ * @throws Exception\InvalidArgumentException
  */
 function is_host_in_noproxy(string $host, array $noProxyArray): bool
 {
-    if (\strlen($host) === 0) {
-        throw new \InvalidArgumentException('Empty host provided');
-    }
-
-    // Strip port if present.
-    if (\strpos($host, ':')) {
-        $host = \explode($host, ':', 2)[0];
-    }
-
-    foreach ($noProxyArray as $area) {
-        // Always match on wildcards.
-        if ($area === '*') {
-            return true;
-        } elseif (empty($area)) {
-            // Don't match on empty values.
-            continue;
-        } elseif ($area === $host) {
-            // Exact matches.
-            return true;
-        } else {
-            // Special match if the area when prefixed with ".". Remove any
-            // existing leading "." and add a new leading ".".
-            $area = '.' . \ltrim($area, '.');
-            if (\substr($host, -(\strlen($area))) === $area) {
-                return true;
-            }
-        }
-    }
-
-    return false;
+    return Utils::isHostInNoProxy($host, $noProxyArray);
 }
 
 /**
@@ -269,14 +125,7 @@ function is_host_in_noproxy(string $host, array $noProxyArray): bool
  */
 function json_decode(string $json, bool $assoc = false, int $depth = 512, int $options = 0)
 {
-    $data = \json_decode($json, $assoc, $depth, $options);
-    if (JSON_ERROR_NONE !== \json_last_error()) {
-        throw new Exception\InvalidArgumentException(
-            'json_decode error: ' . \json_last_error_msg()
-        );
-    }
-
-    return $data;
+    return Utils::jsonDecode($json, $assoc, $depth, $options);
 }
 
 /**
@@ -292,82 +141,5 @@ function json_decode(string $json, bool $assoc = false, int $depth = 512, int $o
  */
 function json_encode($value, int $options = 0, int $depth = 512): string
 {
-    $json = \json_encode($value, $options, $depth);
-    if (JSON_ERROR_NONE !== \json_last_error()) {
-        throw new Exception\InvalidArgumentException(
-            'json_encode error: ' . \json_last_error_msg()
-        );
-    }
-
-    return $json;
-}
-
-/**
- * Wrapper for the hrtime() or microtime() functions
- * (depending on the PHP version, one of the two is used)
- *
- * @return float|mixed UNIX timestamp
- *
- * @internal
- */
-function _current_time()
-{
-    return \function_exists('hrtime') ? \hrtime(true) / 1e9 : \microtime(true);
-}
-
-/**
- * @internal
- */
-function _idn_uri_convert(UriInterface $uri, int $options = 0): UriInterface
-{
-    if ($uri->getHost()) {
-        $idnaVariant = defined('INTL_IDNA_VARIANT_UTS46') ? INTL_IDNA_VARIANT_UTS46 : 0;
-        $asciiHost = $idnaVariant === 0
-            ? idn_to_ascii($uri->getHost(), $options)
-            : idn_to_ascii($uri->getHost(), $options, $idnaVariant, $info);
-        if ($asciiHost === false) {
-            $errorBitSet = isset($info['errors']) ? $info['errors'] : 0;
-
-            $errorConstants = array_filter(array_keys(get_defined_constants()), function ($name) {
-                return substr($name, 0, 11) === 'IDNA_ERROR_';
-            });
-
-            $errors = [];
-            foreach ($errorConstants as $errorConstant) {
-                if ($errorBitSet & constant($errorConstant)) {
-                    $errors[] = $errorConstant;
-                }
-            }
-
-            $errorMessage = 'IDN conversion failed';
-            if ($errors) {
-                $errorMessage .= ' (errors: ' . implode(', ', $errors) . ')';
-            }
-
-            throw new InvalidArgumentException($errorMessage);
-        } else {
-            if ($uri->getHost() !== $asciiHost) {
-                // Replace URI only if the ASCII version is different
-                $uri = $uri->withHost($asciiHost);
-            }
-        }
-    }
-
-    return $uri;
-}
-
-/**
- * @internal
- */
-function _getenv(string $name): ?string
-{
-    if (isset($_SERVER[$name])) {
-        return (string) $_SERVER[$name];
-    }
-
-    if (PHP_SAPI === 'cli' && ($value = \getenv($name)) !== false && $value !== null) {
-        return (string) $value;
-    }
-
-    return null;
+    return Utils::jsonEncode($value, $options, $depth);
 }
