+++ tests/InternalUtilsTest.php
@@ -1,0 +1,25 @@
+<?php
+namespace GuzzleHttp\Test;
+
+use GuzzleHttp\Psr7;
+use GuzzleHttp\Utils;
+use PHPUnit\Framework\TestCase;
+
+class InternalUtilsTest extends TestCase
+{
+    public function testCurrentTime()
+    {
+        self::assertGreaterThan(0, Utils::currentTime());
+    }
+
+    public function testIdnConvert()
+    {
+        if (!extension_loaded('intl')) {
+            self::markTestSkipped('intl PHP extension is not loaded');
+        }
+
+        $uri = Psr7\uri_for('https://яндекс.рф/images');
+        $uri = Utils::idnUriConvert($uri);
+        self::assertSame('xn--d1acpjx3f.xn--p1ai', $uri->getHost());
+    }
+}
