+++ tests/functionsTest.php
@@ -134,22 +134,6 @@ class FunctionsTest extends TestCase
 
         \GuzzleHttp\json_decode('{{]]');
     }
-
-    public function testCurrentTime()
-    {
-        self::assertGreaterThan(0, GuzzleHttp\_current_time());
-    }
-
-    public function testIdnConvert()
-    {
-        if (!extension_loaded('intl')) {
-            self::markTestSkipped('intl PHP extension is not loaded');
-        }
-
-        $uri = GuzzleHttp\Psr7\uri_for('https://яндекс.рф/images');
-        $uri = GuzzleHttp\_idn_uri_convert($uri);
-        self::assertSame('xn--d1acpjx3f.xn--p1ai', $uri->getHost());
-    }
 }
 
 final class StrClass
